termionix_server

Struct AnsiSelectGraphicRendition

pub struct AnsiSelectGraphicRendition {

    pub intensity: Option<Intensity>,
    pub italic: Option<bool>,
    pub underline: Option<Underline>,
    pub blink: Option<Blink>,
    pub reverse: Option<bool>,
    pub hidden: Option<bool>,
    pub strike: Option<bool>,
    pub script: Option<Script>,
    pub ideogram: Option<Ideogram>,
    pub font: Option<Font>,
    pub foreground: Option<Color>,
    pub background: Option<Color>,
    pub unknown: Vec<SGRParameter>,

}

Represents a text style with various formatting attributes and colors for terminal output.

Style encapsulates all ANSI terminal styling properties, including text attributes (bold, italic, underline, etc.) and foreground/background colors. It provides methods to write ANSI escape sequences to apply these styles in terminal environments.

Examples

Basic Usage

use termionix_ansicodec::{Style, Color, ColorMode, Intensity, AnsiConfig};
use std::fmt::Write;

let config = AnsiConfig::enabled();
let mut style = Style::default();
style.intensity = Some(Intensity::Bold);
style.foreground = Some(Color::Red);

let mut output = String::new();
style.write_style(&mut output, Some(&config)).unwrap();
write!(&mut output, "Bold Red Text").unwrap();
Style::write_reset(&mut output, Some(&config)).unwrap();

Complex Styling

use termionix_ansicodec::{Underline,Style, Color, ColorMode, Intensity};

let style = Style {
    intensity: Some(Intensity::Bold),
    italic: Some(true),
    underline: Some(Underline::Single),
    foreground: Some(Color::RGB(255, 100, 50)),
    background: Some(Color::Blue),
    ..Default::default()
};

ANSI Escape Codes

The Style struct generates ANSI escape sequences in the format \x1b[<codes>m, where <codes> is a semicolon-separated list of numeric codes:

Text attributes: 1 (bold), 2 (dim), 3 (italic), 4 (underline), 5 (blink), 7 (reverse), 8 (hidden), 9 (strikethrough)

Foreground colors: 30-37 (basic), 38;5;<n> (256-color), 38;2;<r>;<g>;<b> (RGB)

Background colors: 40-47 (basic), 48;5;<n> (256-color), 48;2;<r>;<g>;<b> (RGB)

Reset: 0 (clear all styles)

Color Mode Support

Style rendering respects the ColorMode setting:

ColorMode::None: No ANSI codes are written

ColorMode::Basic: Basic 8-color support (codes 30-37, 40-47)

ColorMode::FixedColor: 256-color support

ColorMode::TrueColor: Full 24-bit RGB color support

Default Values

The default style has no formatting attributes enabled and no colors set:

use termionix_ansicodec::{Color, Style};

let style = Style::default();
assert_eq!(style.intensity, None);
assert_eq!(style.italic, None);
assert_eq!(style.foreground, None);
assert_eq!(style.background, None);

Fields
intensity: Option<Intensity>

Intensity of Text

Code	Description

1	Bold

2	Dim

22	Normal

See: Intensity

italic: Option<bool>

Whether this style is italic or oblique.

Code	Description

3	Italic

23	Oblique

underline: Option<Underline>

Whether this style is underlined. (Code: 4)

Code	Description

4	Underline

21	Double Underline

24	No Underline

See: Intensity

blink: Option<Blink>

Whether this style is blinking.

Code	Description

5	Blink

6	Rapid Blink

25	No Blink

reverse: Option<bool>

Whether this style has reverse colors.

Code	Description

7	Reverse/Inverse

27	Normal

hidden: Option<bool>

Whether this style is hidden.

Code	Description

8	Hidden/Concealed

28	Normal/Revealed

strike: Option<bool>

Whether this style is struckthrough.

Code	Description

9	Strikethrough

29	Normal

script: Option<Script>

Weather this style is super or subscript.

Code	Description

73	Superscript

74	Subscript

75	Normal

ideogram: Option<Ideogram>

Weather this style is super or subscript.

Code	Description

60	Underline

61	DoubleUnderline

62	Overline

63	DoubleOverline

64	StressMarking

65	NoIdeogramAttributes

font: Option<Font>

The font to use for the text.

Code	Description

10	Primary Font

11	Alternate Font 1

12	Alternate Font 2

13	Alternate Font 3

14	Alternate Font 4

15	Alternate Font 5

16	Alternate Font 6

17	Alternate Font 7

18	Alternate Font 8

19	Alternate Font 9

20	Fraktur Font

foreground: Option<Color>

The foreground color of the text.

Code	Description

30	Black

31	Red

32	Green

33	Yellow

34	Blue

35	Magenta

36	Cyan

37	White

38;5;n	8-bit Fixed

38;2;r;g;b	24-bit RGB

90	Bright Black

91	Bright Red

92	Bright Green

93	Bright Yellow

94	Bright Blue

95	Bright Magenta

96	Bright Cyan

97	Bright White

background: Option<Color>

The background color of the text.

Code	Description

Code	Description

–––––––	––––––––

40	Black

41	Red

42	Green

43	Yellow

44	Blue

45	Magenta

46	Cyan

47	White

48;5;n	8 bit Fixed

48;2;r;g;b	24-bit RGB

100	Bright Black

101	Bright Red

102	Bright Green

103	Bright Yellow

104	Bright Blue

105	Bright Magenta

106	Bright Cyan

107	Bright White

unknown: Vec<SGRParameter>

Remaining SGR Bytes

Implementations

impl AnsiSelectGraphicRendition

pub fn len(&self, _color_mode: Option<ColorMode>) -> usize

Length of Style Control TODO: Use color_mode parameter

pub fn encode<T>( &self, dst: &mut T, color_mode: Option<ColorMode>, ) -> Result<usize, AnsiCodecError>
where T: BufMut,

pub fn write_reset<W>(writer: &mut W) -> Result<(), Error>
where W: Write,

Resets terminal color attributes by writing an ANSI reset escape code (\x1b[0m) to the provided writer.

Parameters

mode: A reference to a ColorMode object that determines if ANSI color codes should be used.

writer: A mutable reference to a writer implementing the std::fmt::Write trait, where the reset escape code will be written if applicable.

Returns

std::fmt::Result: Returns Ok(()) if successful, or an error if writing to the writer fails.

Behavior

If the provided ColorMode does not support ANSI (i.e., mode.is_ansi() is false), the function does nothing and exits early with Ok(()).

If ANSI color is enabled, the function writes the ANSI reset escape code (\x1b[0m) to the given writer.

Example

use termionix_ansicodec::{AnsiConfig, ColorMode, Style};
use std::fmt::Write;
use std::io::BufWriter;

let config = AnsiConfig::enabled();
let mut output = String::new();

Style::write_reset(&mut output, Some(&config)).unwrap();
assert_eq!(output, "\x1b[0m"); // Check that the ANSI reset code is written

pub fn write<W>( &self, writer: &mut W, color_mode: Option<ColorMode>, ) -> Result<usize, Error>
where W: Write,

Writes ANSI style escape codes to the given writer based on the style attributes and colors defined in the instance. This method is used to apply text formatting when ANSI color output is enabled.

Parameters

mode: A reference to a ColorMode instance that determines if ANSI colors are supported. If ANSI is not supported, no styling is applied, and the function returns early.

writer: A mutable reference to a type that implements the std::fmt::Write trait, where the ANSI escape sequences will be written.

Returns

std::fmt::Result: Returns Ok(()) if the escape codes are successfully written or if no styling/formatting is needed. Returns an error if writing to the writer fails.

Behavior

When ANSI coloring is disabled (e.g., ColorMode::is_ansi is false), the method returns immediately without writing any styles.

For each enabled text attribute (e.g., bold, italic, underline, etc.), the corresponding escape code is added to a sequence of codes.

If foreground or background colors are specified, they are included in the ANSI escape sequence. Supports standard, fixed 256-color, and RGB color codes.

If no attributes or colors are specified, no output is written, and the function exits normally.

Example

use termionix_ansicodec::{Style, Color, ColorMode, Intensity, Underline, Blink, AnsiConfig};
use std::fmt::Write;

let config = AnsiConfig::enabled();
let style = Style {
    intensity: Some(Intensity::Bold),
    italic: None,
    underline: Some(Underline::Single),
    blink: Some(Blink::Off),
    reverse: None,
    hidden: None,
    strike: None,
    font: None,
    script: None,
    ideogram: None,
    foreground: Some(Color::Red),
    background: Some(Color::RGB(10, 20, 30)),
    unknown: Vec::new(),
};

let mut output = String::new();
style.write_style(&mut output, Some(&config)).unwrap();
assert_eq!(output, "\x1b[1;4;25;31;48;2;10;20;30m");

Notes

This method assumes an implementation of a Style struct with text attributes such as bold or underline, as well as foreground and background fields for colors.

The Color enum is expected to support standard colors, fixed 256-color values, and custom RGB values.

pub fn write_str<W>( &self, writer: &mut W, color_mode: Option<ColorMode>, ) -> Result<(), Error>
where W: Write,

pub fn parse(params: &[u8]) -> AnsiSelectGraphicRendition

Parses SGR (Select Graphic Rendition) parameters into a Style struct.

This function takes a slice of SGR parameter codes (the numeric values between ESC[ and m in ANSI escape sequences) and converts them into a Style struct with the appropriate formatting attributes and colors.

Arguments

params - A slice of u8 values representing SGR codes (e.g., [1, 31] for bold red)

Returns

A Style struct with the parsed attributes applied. If the input is empty or contains only a reset code (0), returns a default style.

SGR Code Support

Text Attributes

0 - Reset all attributes to default

1 - Bold

2 - Dim

3 - Italic

4 - Underline

5 - Slow blink

6 - Rapid blink

7 - Reverse video

8 - Hidden/concealed

9 - Strikethrough

21 - Double underline

22 - Normal intensity (neither bold nor dim)

23 - Not italic

24 - Not underlined

25 - Not blinking

27 - Not reversed

28 - Not hidden

29 - Not strikethrough

Foreground Colors (Basic)

30-37 - Black, Red, Green, Yellow, Blue, Purple, Cyan, White

39 - Default foreground color

90-97 - Bright variants of basic colors

Background Colors (Basic)

40-47 - Black, Red, Green, Yellow, Blue, Purple, Cyan, White

49 - Default background color

100-107 - Bright variants of basic colors

Extended Colors

38;5;n - Set foreground to 256-color palette color n

38;2;r;g;b - Set foreground to RGB color

48;5;n - Set background to 256-color palette color n

48;2;r;g;b - Set background to RGB color

Fonts

10-19 - Select font (primary or alternate fonts 1-9)

20 - Fraktur font

Notes

Unknown or unsupported SGR codes are collected in style.unknown

Extended color sequences (38/48 with 5 or 2) consume multiple parameters

If extended color sequences are incomplete, the codes are stored as unknown

Reset code (0) clears all attributes and returns a default style