termionix_server

Trait ServerHandler

pub trait ServerHandler:
    Send
    + Sync
    + 'static {
    // Provided methods
    fn on_connect<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_event<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
        _event: TerminalEvent,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_option_changed<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
        _option: TelnetOption,
        _enabled: bool,
        _local: bool,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_subnegotiation<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
        _subneg: TelnetArgument,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_error<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
        _error: TelnetError,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_timeout<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_idle_timeout<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn on_disconnect<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _id: ConnectionId,
        _conn: &'life1 TelnetConnection,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Server event handler trait

Implement this trait to handle events from the Telnet server. All methods are async and have default implementations that do nothing.

Example

use termionix_server::{ServerHandler, ConnectionId, TelnetConnection};
use termionix_terminal::TerminalEvent;
use async_trait::async_trait;

struct MyHandler;

#[async_trait]
impl ServerHandler for MyHandler {
    async fn on_event(
        &self,
        id: ConnectionId,
        conn: &TelnetConnection,
        event: TerminalEvent,
    ) {
        // Handle the event
    }
}

Provided Methods

fn on_connect<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a new connection is established

This is called after the connection is fully initialized and before any events are processed.

fn on_event<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, _event: TerminalEvent, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a terminal event is received

This is the main event processing method. It will be called for every event received from the client.

fn on_option_changed<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, _option: TelnetOption, _enabled: bool, _local: bool, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a Telnet option state changes

This is called when an option negotiation completes successfully, whether the option is being enabled or disabled.

Arguments

id - The connection ID

conn - The connection handle

option - The Telnet option that changed

enabled - true if the option was enabled, false if disabled

local - true if the option changed locally, false if remotely

fn on_subnegotiation<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, _subneg: TelnetArgument, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a Telnet subnegotiation is received

This is called when a complete subnegotiation sequence is received from the client. Subnegotiations provide additional parameters for negotiated options (e.g., window size for NAWS, terminal type for TTYPE).

Arguments

id - The connection ID

conn - The connection handle

subneg - The parsed subnegotiation data

fn on_error<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, _error: TelnetError, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when an error occurs on a connection

This is called when an error occurs during event processing. The connection will be closed after this method returns.

fn on_timeout<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a read operation times out

This is called when no data is received within the configured read timeout. The connection will be closed after this method returns.

fn on_idle_timeout<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a connection is idle for too long

This is called when there has been no activity on the connection for the configured idle timeout. The connection will be closed after this method returns.

fn on_disconnect<'life0, 'life1, 'async_trait>( &'life0 self, _id: ConnectionId, _conn: &'life1 TelnetConnection, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called when a connection is disconnected

This is called when the connection is closed, either by the client, the server, or due to an error.